requirements.txt
django==1.11
djangorestframework==3.6.2
pytest-django==3.1.2
coreapi==2.3.0
